+++ asv_bench/benchmarks/categoricals.py
@@ -220,25 +220,6 @@ class Rank:
         self.s_int_cat_ordered.rank()
 
 
-class IsInCategorical:
-
-    params = ["object", "int64"]
-    param_names = ["dtype"]
-
-    def setup(self, dtype):
-        np.random.seed(1234)
-        n = 5 * 10 ** 5
-        sample_size = 100
-        arr = list(np.random.randint(0, n // 10, size=n))
-        if dtype == "object":
-            arr = [f"s{i:04d}" for i in arr]
-        self.values = np.random.choice(arr, sample_size)
-        self.series = pd.Series(arr).astype("category")
-
-    def time_isin(self, dtype):
-        self.series.isin(self.values)
-
-
 class IsMonotonic:
     def setup(self):
         N = 1000
