+++ asv_bench/benchmarks/series_methods.py
@@ -2,10 +2,7 @@ from datetime import datetime
 
 import numpy as np
 
-from pandas.compat.numpy import np_version_under1p20
-
 from pandas import (
-    Categorical,
     NaT,
     Series,
     date_range,
@@ -28,183 +25,6 @@ class SeriesConstructor:
 
     def time_constructor(self, data):
         Series(data=self.data, index=self.idx)
-
-
-class IsIn:
-
-    params = ["int64", "uint64", "object", "Int64", "boolean", "bool", "datetime64[ns]"]
-    param_names = ["dtype"]
-
-    def setup(self, dtype):
-        N = 10000
-
-        self.mismatched = [NaT.to_timedelta64()] * 2
-
-        if dtype in ["boolean", "bool"]:
-            self.series = Series(np.random.randint(0, 2, N)).astype(dtype)
-            self.values = [True, False]
-
-        elif dtype == "datetime64[ns]":
-            # Note: values here is much larger than non-dt64ns cases
-
-            # dti has length=115777
-            dti = date_range(
-                start=datetime(2015, 10, 26), end=datetime(2016, 1, 1), freq="50s"
-            )
-            self.series = Series(dti)
-            self.values = self.series._values[::3]
-            self.mismatched = [1, 2]
-
-        else:
-            self.series = Series(np.random.randint(1, 10, N)).astype(dtype)
-            self.values = [1, 2]
-
-        self.cat_values = Categorical(self.values)
-
-    def time_isin(self, dtype):
-        self.series.isin(self.values)
-
-    def time_isin_categorical(self, dtype):
-        self.series.isin(self.cat_values)
-
-    def time_isin_empty(self, dtype):
-        self.series.isin([])
-
-    def time_isin_mismatched_dtype(self, dtype):
-        self.series.isin(self.mismatched)
-
-
-class IsInFloat64:
-
-    params = [
-        [np.float64, "Float64"],
-        ["many_different_values", "few_different_values", "only_nans_values"],
-    ]
-    param_names = ["dtype", "title"]
-
-    def setup(self, dtype, title):
-        N_many = 10 ** 5
-        N_few = 10 ** 6
-        self.series = Series([1, 2], dtype=dtype)
-
-        if title == "many_different_values":
-            # runtime is dominated by creation of the lookup-table
-            self.values = np.arange(N_many, dtype=np.float64)
-        elif title == "few_different_values":
-            # runtime is dominated by creation of the lookup-table
-            self.values = np.zeros(N_few, dtype=np.float64)
-        elif title == "only_nans_values":
-            # runtime is dominated by creation of the lookup-table
-            self.values = np.full(N_few, np.nan, dtype=np.float64)
-        else:
-            raise ValueError(title)
-
-    def time_isin(self):
-        self.series.isin(self.values)
-
-
-class IsInForObjects:
-    """
-    A subset of the cartesian product of cases have special motivations:
-
-    "nans" x "nans"
-        if nan-objects are different objects,
-        this has the potential to trigger O(n^2) running time
-
-    "short" x "long"
-        running time dominated by the preprocessing
-
-    "long" x "short"
-        running time dominated by look-up
-
-    "long" x "long"
-        no dominating part
-
-    "long_floats" x "long_floats"
-        because of nans floats are special
-        no dominating part
-
-    """
-
-    variants = ["nans", "short", "long", "long_floats"]
-
-    params = [variants, variants]
-    param_names = ["series_type", "vals_type"]
-
-    def setup(self, series_type, vals_type):
-        if series_type == "nans":
-            self.series = Series(np.full(10 ** 4, np.nan)).astype(object)
-        elif series_type == "short":
-            self.series = Series(np.arange(2)).astype(object)
-        elif series_type == "long":
-            self.series = Series(np.arange(10 ** 5)).astype(object)
-        elif series_type == "long_floats":
-            self.series = Series(np.arange(10 ** 5, dtype=np.float_)).astype(object)
-
-        if vals_type == "nans":
-            self.values = np.full(10 ** 4, np.nan).astype(object)
-        elif vals_type == "short":
-            self.values = np.arange(2).astype(object)
-        elif vals_type == "long":
-            self.values = np.arange(10 ** 5).astype(object)
-        elif vals_type == "long_floats":
-            self.values = np.arange(10 ** 5, dtype=np.float_).astype(object)
-
-    def time_isin(self, series_type, vals_type):
-        self.series.isin(self.values)
-
-
-class IsInLongSeriesLookUpDominates:
-    params = [
-        ["int64", "int32", "float64", "float32", "object", "Int64", "Float64"],
-        [5, 1000],
-        ["random_hits", "random_misses", "monotone_hits", "monotone_misses"],
-    ]
-    param_names = ["dtype", "MaxNumber", "series_type"]
-
-    def setup(self, dtype, MaxNumber, series_type):
-        N = 10 ** 7
-
-        if not np_version_under1p20 and dtype in ("Int64", "Float64"):
-            raise NotImplementedError
-
-        if series_type == "random_hits":
-            np.random.seed(42)
-            array = np.random.randint(0, MaxNumber, N)
-        if series_type == "random_misses":
-            np.random.seed(42)
-            array = np.random.randint(0, MaxNumber, N) + MaxNumber
-        if series_type == "monotone_hits":
-            array = np.repeat(np.arange(MaxNumber), N // MaxNumber)
-        if series_type == "monotone_misses":
-            array = np.arange(N) + MaxNumber
-        self.series = Series(array).astype(dtype)
-        self.values = np.arange(MaxNumber).astype(dtype)
-
-    def time_isin(self, dtypes, MaxNumber, series_type):
-        self.series.isin(self.values)
-
-
-class IsInLongSeriesValuesDominate:
-    params = [
-        ["int64", "int32", "float64", "float32", "object", "Int64", "Float64"],
-        ["random", "monotone"],
-    ]
-    param_names = ["dtype", "series_type"]
-
-    def setup(self, dtype, series_type):
-        N = 10 ** 7
-        if series_type == "random":
-            np.random.seed(42)
-            vals = np.random.randint(0, 10 * N, N)
-        if series_type == "monotone":
-            vals = np.arange(N)
-        self.values = vals.astype(dtype)
-        M = 10 ** 6 + 1
-        self.series = Series(np.arange(M)).astype(dtype)
-
-    def time_isin(self, dtypes, series_type):
-        self.series.isin(self.values)
 
 
 class NSort:
